+++ CHANGES.md
@@ -3,6 +3,8 @@ unrealeased
 
 - Improve the error message when the identifier is not found
   [\27](https://github.com/tarides/ocaml.nvim/pull/27)
+- Enable custom OCaml LSP server name in the config
+  [\26](https://github.com/tarides/ocaml.nvim/pull/26)
 
 ocaml.nvim 1.0.0
 -----------